Добавлено отображение калорий на DashboardScreen.kt Closes #57

+++ app/src/main/java/com/niked/fatless/ui/component/fatlesshistory/FatLessHistoryModels.kt
@@ -21,7 +21,7 @@ data class NutritionBarModel(
     val proteins: Float,
     val fats: Float,
     val carbs: Float,
-    val totalCalories: Int,
+    val totalCalories: Float,
     val isToday: Boolean,
     val isFuture: Boolean
 )
